requirement.txt
streamlit
openai-whisper
pandas
numpy
torch
torchaudio
tqdm
ffmpeg-python